batch: Move ownership unit selection
Ownership units now have separate value objects, construction helpers,
validation, and rebuild logic inside the batch package.

The unit operation module still owns display-ID selection and filtering.
That makes command selection semantics sit beside unit construction, even
though the selection helpers only consume completed unit records.

This commit continues separating ownership unit responsibilities by moving
display-ID selection and filtering into an ownership unit selection module,
then updating the batch selection and reset flows to import that boundary.

Subsequent commits will update tests and architecture guards for the new
selection owner.
